Move user endpoint to separate file to avoid conflicts
Change db.js in two ways:

 * The function didn't need to be async (it didn't do any async stuff
   inside).

 * It also didn't create the pool at all other than the first time, so
   just create it upfront.

While at it, use the simpler pool.query() rather than getConnection(),
since we only want to do a single query. That way we don't need to call
connection.release() later.

Move dotenv.config() call to the very beginning of index.js instead of
in db.js. That way we don't rely on db.js being the first thing imported
to be able to have the env variables.